fix(test): relocate link-navigation tests to correct directory and remove global mock pollution
- Move link-navigation tests from views/infra/__tests__/ to views/router/__tests__/ where the Link component source lives; merge with existing ctrl+click test to cover all three modifier keys (meta, ctrl, shift) in one test
- Remove link-agent entry from mockLogDetails and global telemetry/agent handler from v1-runs.ts; these were test-specific and are already provided inline via server.use() in the test file
- Guard openSpy.mockRestore() with try/finally to prevent spy leaks on intermediate assertion failures
